# internal/repository/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.25.0

package repository

import (
	"database/sql"
	"time"

	"github.com/google/uuid"
)

type EntryLog struct {
	ID          uuid.UUID      `json:"id"`
	PassID      uuid.UUID      `json:"pass_id"`
	GuardUserID uuid.UUID      `json:"guard_user_id"`
	Action      string         `json:"action"`
	ActionAt    time.Time      `json:"action_at"`
	Comment     sql.NullString `json:"comment"`
}

type GuestRequest struct {
	ID             uuid.UUID     `json:"id"`
	ResidentUserID uuid.UUID     `json:"resident_user_id"`
	GuestFullName  string        `json:"guest_full_name"`
	PlateNumber    string        `json:"plate_number"`
	ValidFrom      time.Time     `json:"valid_from"`
	ValidTo        time.Time     `json:"valid_to"`
	Status         string        `json:"status"`
	CreatedAt      time.Time     `json:"created_at"`
	UpdatedAt      time.Time     `json:"updated_at"`
	CreatedBy      uuid.NullUUID `json:"created_by"`
	UpdatedBy      uuid.NullUUID `json:"updated_by"`
	DeletedAt      sql.NullTime  `json:"deleted_at"`
}

type Pass struct {
	ID           uuid.UUID      `json:"id"`
	OwnerUserID  uuid.UUID      `json:"owner_user_id"`
	PlateNumber  string         `json:"plate_number"`
	VehicleBrand sql.NullString `json:"vehicle_brand"`
	VehicleColor sql.NullString `json:"vehicle_color"`
	Status       string         `json:"status"`
	CreatedAt    time.Time      `json:"created_at"`
	UpdatedAt    time.Time      `json:"updated_at"`
	CreatedBy    uuid.NullUUID  `json:"created_by"`
	UpdatedBy    uuid.NullUUID  `json:"updated_by"`
	DeletedAt    sql.NullTime   `json:"deleted_at"`
}

type User struct {
	ID           uuid.UUID      `json:"id"`
	Email        string         `json:"email"`
	PasswordHash string         `json:"password_hash"`
	Role         string         `json:"role"`
	FullName     string         `json:"full_name"`
	PlotNumber   sql.NullString `json:"plot_number"`
	BlockedAt    sql.NullTime   `json:"blocked_at"`
	CreatedAt    time.Time      `json:"created_at"`
	UpdatedAt    time.Time      `json:"updated_at"`
	CreatedBy    uuid.NullUUID  `json:"created_by"`
	UpdatedBy    uuid.NullUUID  `json:"updated_by"`
	DeletedAt    sql.NullTime   `json:"deleted_at"`
}